Update form type mapping function so it works with construction
