+++ mbs_results/staging/back_data.py
@@ -2,7 +2,7 @@ import warnings
 
 import pandas as pd
 
-from mbs_results.staging.data_cleaning import enforce_datatypes
+from mbs_results.staging.data_cleaning import enforce_datatypes, create_form_type_spp_column
 from mbs_results.utilities.inputs import read_colon_separated_file, read_csv_wrapper
 from mbs_results.utilities.utils import convert_column_to_datetime
 
@@ -199,11 +199,13 @@ def read_and_process_back_data(config: dict) -> pd.DataFrame:
         back_data[type_col].fillna(-999).astype(int).astype(str).map(map_type),
     )
 
-    # This needed here?
-    idbr_to_spp_mapping = config["idbr_to_spp"]
-    back_data[config["form_id_spp"]] = (
-        back_data[config["form_id_idbr"]].astype(str).map(idbr_to_spp_mapping)
-    )
-    back_data["cellnumber"] = back_data["cell_no"]
+    # TO-DO: Refactor this so construction pipeline doesn't get an unnecessary warning
+    if "idbr_to_spp" in config:
+      create_form_type_spp_column(back_data, config)
+    else:
+      logger.info("""idbr_to_spp column not specified in config, so skipping create_form_type_spp_column. 
+                  You can safely ignore this if this was intentional, i.e. running the construction pipeline.""")
 
+    back_data["cellnumber"] = back_data["cell_no"]
+      
     return back_data
