+++ mbs_results/staging/data_cleaning.py
@@ -10,6 +10,29 @@ from mbs_results.utilities.validation_checks import (  # validate_manual_constru
     validate_indices,
 )
 
+def create_form_type_spp_column(
+    df: pd.DataFrame, config: dict
+) -> pd.DataFrame:
+    """
+    maps IDBR form types to SPP and creates column named "form_type_spp"
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        dataframe containing the data (contributors or backdata)
+    config : dict
+        main pipeline config containing "idbr_to_spp" mapping
+
+    Returns
+    -------
+    pd.DataFrame
+        df dataframe with "form_type_spp" column added
+    """
+    idbr_to_spp_mapping = config["idbr_to_spp"]
+    df[config["form_id_spp"]] = df[config["form_id_idbr"]].astype(str).map(
+        idbr_to_spp_mapping
+    )
+    return df
 
 def filter_responses(df, reference, period, last_update):
     """
